change user id type from int to unsigned long int:

+++ game-logic/User.cpp
@@ -4,9 +4,9 @@
 #include <algorithm>
 
 // Server should generate a unique id upon User construction
-User::User(int id, const std::string& name) : id(id), name(name) {}
+User::User(unsigned long int id, const std::string& name) : id(id), name(name) {}
 
-int User::getId() const {
+unsigned long int User::getId() const {
     return id;
 }
